+++ vnstock/api/trading.py
@@ -39,8 +39,8 @@ class Trading(BaseAdapter):
         self.show_log = show_log
         
         # Validate to accept vci, tcbs as source
-        if source.lower() not in ["kbs", "vci", "tcbs"]:
-            raise ValueError("Lớp Trading chỉ nhận giá trị tham số source là 'VCI' hoặc 'TCBS'.")
+        if source.lower() not in ["kbs", "vci"]:
+            raise ValueError("Lớp Trading chỉ nhận giá trị tham số source là 'VCI' hoặc 'KBS'.")
         
         super().__init__(
             source=source,